ui-test-cleanup - move init module out of _lib/pom into _lib/init - because it is not a pom - it uses pom.
